Cleaned Version of Package
Cleaned Version of Package, now called EffIntThy= Effective Intersection Theory

Added methods called ChowClass and projectiveDegree. projectiveDegree needs a wrapper function (or maybe several) , but I already wrote a wrapper for ChowClass

I now use a hash table called Info or ShareInfo, to move needed values between functions. This could be converted to an object, though I am not sure it is helpful to the user to have an object, or at least it shouldn't be mandatory for the user to make one. I.e. I would like to preserve the behavior 

s=Segre(I,J) 
where I and J are ideals and the returned value s is a a class in an explicit quotient ring of the form Z[h_1..h_m]/(h_1^n_1,...,h_m^n_m), we could also overload the function to take other inputs and add options for other return types, but I think a simple give ideal get ring element is best as the "standard" way to interact with the function.